pyproject.toml
[project]
name = "openai-agent-sdk-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.13.5",
    "playwright>=1.58.0",
]
